Pen2 draw function work
The pen2 without XRGrapInteractive and ridgetbody work perfectly, and it draw on the surface of the side it first touch.
Now a line is able to draw on the surface from the touching point till the pen move outside the boundare of the drawing surface.

What it need now:
Primary is to make it draw perpendicular line on the substance drawing surface.
Secondary: Make sure the pen is drawing on the right surface depending on the pens current normal  pointing direction