Much change
A lot had changed since the last push up and a lot of them are meaning less.

Now this only working thing is that the saw is locked on the surface of the touching substance and move with the change of the hand position, but is not what it shall.

On the next puch up, if it work, then the saw will be able to move based on the dot product of the grabbing hands motion on the direciton it forward vector currently facing and it up vector will move inverted if the dot product is not 0