Temp fix of ghost plank
The previous slicing method keep createing ghost plank due to the sliced plank object somehow kept the reference from the original object for now I cannot find the code of it.

It now have another slice method that restrict the angle cutting.

For now I will work on intercepting with the sliced plank objects and interception display method to show that the hand and tools touching with the plank will have some way to display they are intercepting